refactor: simplify code and improve readability across multiple modules
- Replaced `map_or` with `is_none_or` for better clarity in condition checks.
- Updated various string manipulation methods to use more concise and efficient alternatives.
- Enhanced error handling by simplifying error mapping in asynchronous functions.
- Improved path handling by utilizing more efficient methods for trimming and splitting strings.
- Cleaned up unnecessary comments and improved overall code organization for better maintainability.

+++ apps/desktop/src-tauri/src/utils/launch_exe.rs
@@ -54,7 +54,7 @@ fn windows_launch(path: &str) -> Result<(), String> {
                 path_w.as_ptr(),
                 ptr::null(),
                 dir_ptr,
-                SW_SHOWNORMAL as i32,
+                SW_SHOWNORMAL,
             );
             if (h as isize) <= 32 {
                 Err(format!(
